Move filter and use FilterRegistrationBean
