+++ service/src/main/java/org/databiosphere/workspacedataservice/service/CacheControlFilter.java
@@ -1,13 +1,9 @@
-package org.databiosphere.workspacedataservice.config;
+package org.databiosphere.workspacedataservice.service;
 
 import jakarta.servlet.*;
-import jakarta.servlet.annotation.WebFilter;
 import jakarta.servlet.http.HttpServletResponse;
 import java.io.IOException;
 
-@WebFilter(
-    asyncSupported = true,
-    urlPatterns = {"/*"})
 public class CacheControlFilter implements Filter {
   @Override
   public void doFilter(ServletRequest request, ServletResponse response, FilterChain chain)
